Update portage tree to 2017-03-19

+++ app-emulation/open-vm-tools/open-vm-tools-10.1.0.ebuild
@@ -1,6 +1,5 @@
 # Copyright 1999-2016 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Id$
 
 EAPI=6
 MODULES_OPTIONAL_USE="modules"
@@ -122,8 +121,10 @@ src_install() {
 	default
 	prune_libtool_files --modules
 
-	rm "${ED%/}"/etc/pam.d/vmtoolsd || die
-	pamd_mimic_system vmtoolsd auth account
+	if use pam; then
+		rm "${ED%/}"/etc/pam.d/vmtoolsd || die
+		pamd_mimic_system vmtoolsd auth account
+	fi
 
 	newinitd "${FILESDIR}/open-vm-tools.initd" vmware-tools
 	newconfd "${FILESDIR}/open-vm-tools.confd" vmware-tools
